requirements.txt
schedule
python-telegram-bot
python-dotenv
psycopg2-binary
flask